fixed runtime issue
